refactor: refactored header.tsx file by adding imported functions
